Redis 分布式锁 的创建

+++ tmall-product/src/main/java/com/ysy/tmall/product/service/impl/CategoryServiceImpl.java
@@ -171,9 +171,9 @@ public class CategoryServiceImpl extends ServiceImpl<CategoryDao, CategoryEntity
                         "else " +
                         "return 0 " +
                         "end";
-                Integer unLock = stringRedisTemplate
+                Long unLock = stringRedisTemplate
                         .execute(new DefaultRedisScript<>(script,
-                                Integer.class),
+                                Long.class),
                                 Arrays.asList("categoryLock"), token);
             }
